pyproject.toml
[tool.poetry]
name = "ashen-fen-zu-ji-qi-ren"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
"discord.py" = "^1.7.3"
"dislash.py" = "^1.4.9"
discord-py-slash-command = "^3.0.3"
slash-cog = "^0.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
